requirements.txt
streamlit>=1.28.0
openai>=1.0.0
lunar-python
pandas>=2.0.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
